Moved *_model.py files to models directory
